♻️ refactor: filter state update, refresh func apply.
